fixed id generator bug

+++ model/Teacher.java
@@ -1,5 +1,7 @@
 package model;
 
+import java.math.BigDecimal;
+
 import util.IdGenerator;
 
 public class Teacher {
@@ -7,10 +9,10 @@ public class Teacher {
 	private final Integer teacherId;
 	private String name;
 	private String subject;
-	private Double salary;
+	private BigDecimal salary;
 
 	// ID generated inside constructor
-	public Teacher(String name, String subject, Double salary) {
+	public Teacher(String name, String subject, BigDecimal salary) {
 
 		this.teacherId = IdGenerator.generateId();
 		this.name = name;
@@ -18,7 +20,7 @@ public class Teacher {
 		this.salary = salary;
 	}
 
-	public Teacher(Integer teacherId, String name, String subject, Double salary) {
+	public Teacher(Integer teacherId, String name, String subject, BigDecimal salary) {
 
 		this.teacherId = teacherId;
 		this.name = name;
@@ -46,11 +48,11 @@ public class Teacher {
 		this.subject = subject;
 	}
 
-	public Double getSalary() {
+	public BigDecimal getSalary() {
 		return salary;
 	}
 
-	public void setSalary(Double salary) {
+	public void setSalary(BigDecimal salary) {
 		this.salary = salary;
 	}
